+++ wemake_python_styleguide/logic/scopes.py
@@ -6,20 +6,12 @@ from typing import ClassVar, DefaultDict, Set, cast
 
 from typing_extensions import final
 
-from wemake_python_styleguide.compat.aliases import AssignNodes, FunctionNodes
 from wemake_python_styleguide.logic.naming import access, name_nodes
 from wemake_python_styleguide.logic.nodes import get_context
-from wemake_python_styleguide.logic.source import node_to_string
 from wemake_python_styleguide.types import ContextNodes
 
 #: That's how we represent scopes that are bound to contexts.
 _ContextStore = DefaultDict[ContextNodes, Set[str]]
-
-#: That's what we expect from `@overload` decorator:
-_overload_exceptions = frozenset(('overload', 'typing.overload'))
-
-#: That's what we expect from `@property` decorator:
-_property_exceptions = frozenset(('property', '.setter'))
 
 
 class _BaseScope(object):
@@ -137,34 +129,3 @@ class OuterScope(_BaseScope):
 def extract_names(node: ast.AST) -> Set[str]:
     """Extracts unique set of names from a given node."""
     return set(name_nodes.get_variables_from_node(node))
-
-
-def is_function_overload(node: ast.AST) -> bool:
-    """Check that function decorated with `typing.overload`."""
-    if isinstance(node, FunctionNodes):
-        for decorator in node.decorator_list:
-            if node_to_string(decorator) in _overload_exceptions:
-                return True
-    return False
-
-
-def is_property_setter(node: ast.AST, _=None):
-    """Check that function decorated with `property.setter`."""
-    if isinstance(node, FunctionNodes):
-        for decorator in node.decorator_list:
-            if node_to_string(decorator) in _property_exceptions:
-                return True
-    return False
-
-
-def is_same_value_reuse(node: ast.AST, names: Set[str]) -> bool:
-    """Checks if the given names are reused by the given node."""
-    if isinstance(node, AssignNodes) and node.value:
-        used_names = {
-            name_node.id
-            for name_node in ast.walk(node.value)
-            if isinstance(name_node, ast.Name)
-        }
-        if not names.difference(used_names):
-            return True
-    return False